refactor auth code to fit project structure
