+++ src/proxy/routes/index.ts
@@ -1,0 +1,6 @@
+export * from './authorizeGetHandler'
+export * from './MCPDeleteHandler'
+export * from './MCPGetHandler'
+export * from './MCPPostHandler'
+export * from './registerClientPostHandler'
+export * from './tokenPostHandler'
